requirements.txt
faiss-cpu
sentence-transformers
numpy
fastapi
uvicorn
jinja2
python-multipart
